third_party/WebKit/LayoutTests/http/tests/csspaint/style-first-letter-pseudo.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/run-after-layout-and-paint.js"></script>
<script src="resources/test-runner-paint-worklet.js"></script>
<script src="resources/generate-paint-style-logging.js"></script>
<style>
    div {
        color: rgb(0, 255, 0);
        line-height: 1px;
    }

    div::first-letter {
        background-image: paint(test);
        color: rgb(255, 0, 0);
        line-height: 2px;
    }
</style>
</head>
<body>
<div>Foo</div>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();

    importPaintWorkletAndTerminateTestAfterAsyncPaint(
            generatePaintStyleLogging([
                'color',
                'line-height',
            ]));
</script>
<p>This tests the style information in the paint callback.</p>
<p>See the devtools console for test output. The console should log:</p>
color: [CSSStyleValue=rgb(255, 0, 0)]<br>
line-height: [CSSSimpleLength=2px]<br>
</body>
</html>
