third_party/WebKit/LayoutTests/http/tests/csspaint/style-first-letter-pseudo-expected.txt

CONSOLE MESSAGE: line 14: color: [CSSStyleValue=rgb(255, 0, 0)]
CONSOLE MESSAGE: line 14: line-height: [CSSSimpleLength=2px]
Foo
This tests the style information in the paint callback.

See the devtools console for test output. The console should log:

color: [CSSStyleValue=rgb(255, 0, 0)]
line-height: [CSSSimpleLength=2px]
